Moved input format classes to other package
