+++ sdks/java/io/hadoop-input-format/pom.xml
@@ -62,10 +62,6 @@
       <artifactId>beam-sdks-java-core</artifactId>
     </dependency>
     <dependency>
-      <groupId>com.fasterxml.jackson.core</groupId>
-      <artifactId>jackson-annotations</artifactId>
-    </dependency>
-    <dependency>
       <groupId>com.google.guava</groupId>
       <artifactId>guava</artifactId>
       <version>${guava.version}</version>
